+++ CHANGELOG.md
@@ -9,6 +9,11 @@
 ### Breaking
 - Update `@types/desmos` peer dependency `1.11` -> `1.12`
 
+### Fixes
+- Decompiler emits `% ...` for single-line notes, and better formatted `/text{ ... }` for multi-line notes
+- Decompiler rejects tables with `[tables are currently unsupported]` instead of `?`
+- Decompiler rejects other expressions with `[unsupported]` instead of `?`
+
 
 <br>